Clean up component hierarchy and follow gesture handler best practices
- Rename SimpleSwipeableTaskCard to SwipeableTaskCard
- Rename SimpleDraggableTaskCard to DraggableTaskCard
- Remove old complex SwipeableTaskCard and DraggableTaskCard components
- Remove custom gesture hooks (useSwipeGesture, useDragGesture)
- Remove hooks directory entirely
- Update component imports and index file
- Fix unnecessary dependency in DraggableTaskCard

The "Simple" versions follow React Native Gesture Handler best practices
better by implementing gestures directly in components using the new
Gesture API, rather than abstracting them into custom hooks.

Final hierarchy: TaskCard -> SwipeableTaskCard -> DraggableTaskCard
