+++ components/TaskList.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { FlatList, StyleSheet, Text, View } from 'react-native';
 import { Task } from '../types/Task';
-import { SimpleDraggableTaskCard } from './SimpleDraggableTaskCard';
+import { DraggableTaskCard } from './DraggableTaskCard';
 
 interface TaskListProps {
   tasks: Task[];
@@ -22,7 +22,7 @@ export function TaskList({ tasks }: TaskListProps) {
       data={tasks}
       keyExtractor={(item) => item.id}
       renderItem={({ item, index }) => (
-        <SimpleDraggableTaskCard 
+        <DraggableTaskCard 
           task={item} 
           index={index} 
           isDragEnabled={true}
