+++ components/index.ts
@@ -2,8 +2,6 @@ export { DraggableTaskCard } from './DraggableTaskCard';
 export { FAB } from './FAB';
 export { InlineTextEdit } from './InlineTextEdit';
 export { LocationPicker } from './LocationPicker';
-export { SimpleDraggableTaskCard } from './SimpleDraggableTaskCard';
-export { SimpleSwipeableTaskCard } from './SimpleSwipeableTaskCard';
 export { SnoozeActionSheet } from './SnoozeActionSheet';
 export { SuccessAnimation } from './SuccessAnimation';
 export { SwipeableTaskCard } from './SwipeableTaskCard';
