--- hooks/index.ts
@@ -1,2 +1,0 @@
-export { useDragGesture, type DragGestureConfig, type DragGestureReturn } from './useDragGesture';
-export { useSwipeGesture, type SwipeAction, type SwipeGestureConfig, type SwipeGestureReturn } from './useSwipeGesture';
